Putting relaxed_constants into algorithms.
